--- src/main/java/chess/Application.java
@@ -1,11 +1,0 @@
-package chess;
-
-import chess.controller.ChessController;
-
-
-public class Application {
-    public static void main(String[] args) {
-        ChessController chessController = new ChessController();
-        chessController.play();
-    }
-}
