--- src/main/java/chess/domain/Player.java
@@ -1,5 +1,0 @@
-package chess.domain;
-
-public enum Player {
-    WHITE, BLACK
-}
