+++ src/main/java/chess/domain/chesspieces/PieceInfo.java
@@ -1,7 +1,7 @@
 package chess.domain.chesspieces;
 
 import chess.Exceptions.IllegalPlayerException;
-import chess.domain.Player;
+import chess.domain.game.Player;
 import chess.domain.position.component.Column;
 import chess.domain.position.component.Row;
