+++ src/main/java/chess/domain/position/component/Row.java
@@ -2,6 +2,7 @@ package chess.domain.position.component;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 
 public enum Row {
     A('a'),
