+++ src/test/java/ChessBoardTest.java
@@ -1,4 +1,4 @@
-mport chess.Exception.NotMoveException;
+import chess.Exception.NotMoveException;
 import chess.domain.chesspiece.King;
 import chess.domain.chesspiece.Pawn;
 import chess.domain.chesspiece.Piece;
@@ -20,7 +20,6 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -30,8 +29,7 @@ public class ChessBoardTest {
     @DisplayName("초기 체스판 개수 확인")
     @Test
     void initChessBoard() {
-        ChessBoard chessBoard = new ChessBoard();
-        chessBoard.create(PieceFactory.create());
+        ChessBoard chessBoard = new ChessBoard(PieceFactory.create());
         int actual = chessBoard.getChessBoard().size();
         int expected = 32;
         assertThat(actual).isEqualTo(expected);
@@ -40,10 +38,7 @@ public class ChessBoardTest {
     @DisplayName("(예외) 같은 위치로 이동")
     @Test
     void moveSamePosition() {
-        ChessBoard chessBoard = new ChessBoard();
-        chessBoard.create(PieceFactory.create());
-    void 같은_위치로_이동했을때() {
-        ChessBoard chessBoard = new ChessBoard();
+        ChessBoard chessBoard = new ChessBoard(PieceFactory.create());
         Position position = Positions.of("a1");
         assertThatThrownBy(() -> chessBoard.move(position, position))
                 .isInstanceOf(NotMoveException.class);
@@ -53,8 +48,7 @@ public class ChessBoardTest {
     @ParameterizedTest
     @EnumSource(value = Player.class)
     void createStatusTest(Player player) {
-        ChessBoard chessBoard = new ChessBoard();
-        chessBoard.create(PieceFactory.create());
+        ChessBoard chessBoard = new ChessBoard(PieceFactory.create());
         Status result = chessBoard.createStatus(player);
         double actual = result.getScore();
         double expected = 38;
@@ -65,7 +59,7 @@ public class ChessBoardTest {
     @ParameterizedTest
     @MethodSource("generatePositions3")
     void existPawnNumberTest(List<Piece> columnLine, Player player, int exptectd) {
-        ChessBoard chessBoard = new ChessBoard();
+        ChessBoard chessBoard = new ChessBoard(PieceFactory.create());
         assertThat(chessBoard.getPawnCountPerStage(columnLine, player)).isEqualTo(exptectd);
     }
